refactor: update path_helpers imports to shared.paths
Update all imports from vibe3.services.path_helpers to
vibe3.services.shared.paths after migration.

Files updated:
- commands/handoff_read.py
- commands/handoff_render.py
- roles/run_helpers.py
- services/flow_consistency_check.py
- services/handoff_service.py
- services/handoff_storage.py
- services/spec_ref_service.py
- services/task_show_service.py
- ui/flow_ui.py
- ui/flow_ui_primitives.py
- ui/flow_ui_timeline.py
- ui/task_ui.py

Related: #2126
