+++ src/vibe3/roles/run_helpers.py
@@ -194,7 +194,7 @@ def ensure_plan_file_exists(
         return
     if Path(plan_file).is_absolute() and Path(plan_file).exists():
         return
-    from vibe3.services.path_helpers import resolve_handoff_target
+    from vibe3.services.shared.paths import resolve_handoff_target
 
     try:
         resolve_handoff_target(plan_file, branch=branch)
