+++ src/vibe3/services/handoff_storage.py
@@ -7,7 +7,7 @@ from loguru import logger
 
 from vibe3.clients.git_client import GitClient
 from vibe3.services.git_path_client import GitPathProtocol
-from vibe3.services.path_helpers import get_git_common_dir, normalize_ref_path
+from vibe3.services.shared.paths import get_git_common_dir, normalize_ref_path
 from vibe3.utils.git_helpers import get_branch_handoff_dir
