+++ src/vibe3/services/task_show_service.py
@@ -15,8 +15,8 @@ from vibe3.models.flow import FlowStatusResponse
 from vibe3.models.pr import PRResponse
 from vibe3.services.artifact_parser import ArtifactParser
 from vibe3.services.flow_service import FlowService
-from vibe3.services.path_helpers import resolve_ref_path
 from vibe3.services.pr_service import PRService
+from vibe3.services.shared.paths import resolve_ref_path
 from vibe3.utils.comment_utils import is_human_comment
