+++ src/vibe3/ui/flow_ui_timeline.py
@@ -1,7 +1,7 @@
 """Flow UI timeline rendering components."""
 
 from vibe3.models.flow import FlowEvent, FlowStatusResponse
-from vibe3.services.path_helpers import check_ref_exists, ref_to_handoff_cmd
+from vibe3.services.shared.paths import check_ref_exists, ref_to_handoff_cmd
 from vibe3.ui.console import console
 from vibe3.ui.flow_ui_primitives import display_actor, kv, status_text
