+++ src/vibe3/ui/task_ui.py
@@ -6,7 +6,7 @@ from dataclasses import asdict
 from types import ModuleType
 from typing import TYPE_CHECKING
 
-from vibe3.services.path_helpers import ref_to_handoff_cmd
+from vibe3.services.shared.paths import ref_to_handoff_cmd
 from vibe3.ui.console import console
 from vibe3.ui.flow_ui_primitives import resolve_ref_path
 from vibe3.utils.constants import AUTOMATED_MARKERS
